requirements.txt
openai
trafilatura
tqdm
webvtt-py
cardet 
